python-data-engineer/week5/exercises/neo4j_graph.py
"""
Neo4j Graph Database Operations
-----------------------------

TARGET:
1. Graph Structure
   - Node management
   - Relationship types
   - Property schemas

2. Query Operations
   - Cypher patterns
   - Path finding
   - Pattern matching

3. Graph Algorithms
   - Centrality
   - Community detection
   - Similarity metrics

IMPLEMENTATION APPROACH:
-----------------------

1. Graph Management:
   - Node Operations
     * Label organization
     * Property indexing
     * Constraint handling
   - Relationship Types
     * Direction handling
     * Weight properties
     * Temporal aspects
   - Schema Design
     * Property structure
     * Index strategy
     * Constraint rules

2. Query Patterns:
   - Path Finding
     * Shortest paths
     * All paths
     * Weighted paths
   - Pattern Matching
     * Variable length
     * Complex conditions
     * Aggregations
   - Graph Projections
     * Subgraph extraction
     * View creation
     * Memory handling

3. Algorithm Usage:
   - Centrality Metrics
     * PageRank
     * Betweenness
     * Closeness
   - Community Detection
     * Louvain
     * Label propagation
     * Triangle counting
   - Similarity Calculations
     * Node similarity
     * Relationship overlap
     * Path similarity

PERFORMANCE METRICS:
------------------
1. Query Performance
   - Response time
   - Memory usage
   - Cache hits

2. Algorithm Efficiency
   - Execution time
   - Resource usage
   - Result quality

3. Storage Impact
   - Node count
   - Relationship density
   - Property size

Example Usage:
-------------
graph = Neo4jGraph('bolt://localhost:7687')

# Create nodes and relationships
graph.create_node(
    'Person',
    {'name': 'John', 'age': 30}
)
graph.create_relationship(
    'Person', 'KNOWS', 'Person',
    {'since': '2023'}
)

# Run graph algorithm
result = graph.run_algorithm(
    'pagerank',
    {'nodeLabel': 'Person'}
)
"""

from neo4j import GraphDatabase
import logging
from typing import Dict, List, Any
from datetime import datetime
import time

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class Neo4jGraph:
    """Neo4j graph database handler."""
    
    def __init__(
        self,
        uri: str = "bolt://localhost:7687",
        user: str = "neo4j",
        password: str = "password"
    ):
        """Initialize Neo4j connection."""
        self.driver = GraphDatabase.driver(
            uri,
            auth=(user, password)
        )
        logger.info(f"Connected to Neo4j: {uri}")
    
    def create_node(
        self,
        label: str,
        properties: Dict,
        constraints: List[str] = None
    ):
        """Create node with label and properties."""
        try:
            with self.driver.session() as session:
                # Create constraints if needed
                if constraints:
                    for prop in constraints:
                        session.run(f"""
                            CREATE CONSTRAINT IF NOT EXISTS
                            FOR (n:{label})
                            REQUIRE n.{prop} IS UNIQUE
                        """)
                
                # Create node
                result = session.run(f"""
                    CREATE (n:{label} $props)
                    RETURN n
                """, props=properties)
                
                node = result.single()
                logger.info(f"Created node: {label}")
                return node
                
        except Exception as e:
            logger.error(f"Node creation error: {e}")
            raise
    
    def create_relationship(
        self,
        from_label: str,
        rel_type: str,
        to_label: str,
        properties: Dict = None,
        where: Dict = None
    ):
        """Create relationship between nodes."""
        try:
            with self.driver.session() as session:
                # Build query
                query = f"""
                    MATCH (a:{from_label}), (b:{to_label})
                """
                
                if where:
                    conditions = [
                        f"a.{k} = ${k}"
                        for k in where.keys()
                    ]
                    query += f"WHERE {' AND '.join(conditions)}"
                
                query += f"""
                    CREATE (a)-[r:{rel_type}]->(b)
                    {' SET r = $props' if properties else ''}
                    RETURN r
                """
                
                # Execute query
                result = session.run(
                    query,
                    props=properties,
                    **where if where else {}
                )
                
                rel = result.single()
                logger.info(f"Created relationship: {rel_type}")
                return rel
                
        except Exception as e:
            logger.error(f"Relationship error: {e}")
            raise
    
    def run_algorithm(
        self,
        algorithm: str,
        params: Dict = None
    ):
        """Execute graph algorithm."""
        try:
            with self.driver.session() as session:
                start_time = time.time()
                
                # Algorithm selection
                if algorithm == 'pagerank':
                    query = """
                        CALL gds.pageRank.stream($nodeLabel)
                        YIELD nodeId, score
                        RETURN gds.util.asNode(nodeId).name as name,
                               score
                        ORDER BY score DESC
                    """
                
                elif algorithm == 'louvain':
                    query = """
                        CALL gds.louvain.stream($nodeLabel)
                        YIELD nodeId, communityId
                        RETURN gds.util.asNode(nodeId).name as name,
                               communityId
                        ORDER BY communityId
                    """
                
                else:
                    raise ValueError(f"Unknown algorithm: {algorithm}")
                
                # Execute algorithm
                result = session.run(query, params or {})
                results = list(result)
                
                duration = time.time() - start_time
                
                metrics = {
                    'algorithm': algorithm,
                    'results': len(results),
                    'duration': f"{duration:.6f}s"
                }
                
                logger.info(f"Algorithm metrics: {metrics}")
                return results
                
        except Exception as e:
            logger.error(f"Algorithm error: {e}")
            raise
    
    def query_paths(
        self,
        start_label: str,
        end_label: str,
        rel_type: str = None,
        max_depth: int = 3,
        where: Dict = None
    ):
        """Find paths between nodes."""
        try:
            with self.driver.session() as session:
                # Build query
                query = f"""
                    MATCH path = (start:{start_label})-
                    [{f':{rel_type}' if rel_type else ''} *1..{max_depth}]-
                    (end:{end_label})
                """
                
                if where:
                    conditions = [
                        f"start.{k} = ${k}"
                        for k in where.keys()
                    ]
                    query += f"WHERE {' AND '.join(conditions)}"
                
                query += """
                    RETURN path,
                           length(path) as length,
                           [n IN nodes(path) | n.name] as nodes
                    ORDER BY length
                """
                
                # Execute query
                start_time = time.time()
                result = session.run(
                    query,
                    where or {}
                )
                
                paths = list(result)
                duration = time.time() - start_time
                
                metrics = {
                    'paths_found': len(paths),
                    'max_length': max(
                        p['length'] for p in paths
                    ) if paths else 0,
                    'duration': f"{duration:.6f}s"
                }
                
                logger.info(f"Path query metrics: {metrics}")
                return paths
                
        except Exception as e:
            logger.error(f"Path query error: {e}")
            raise
    
    def monitor_stats(self):
        """Monitor Neo4j statistics."""
        try:
            with self.driver.session() as session:
                stats = {
                    'nodes': session.run("""
                        MATCH (n)
                        RETURN count(n) as count
                    """).single()['count'],
                    
                    'relationships': session.run("""
                        MATCH ()-[r]->()
                        RETURN count(r) as count
                    """).single()['count'],
                    
                    'labels': session.run("""
                        CALL db.labels()
                        YIELD label
                        RETURN collect(label) as labels
                    """).single()['labels'],
                    
                    'types': session.run("""
                        CALL db.relationshipTypes()
                        YIELD relationshipType
                        RETURN collect(relationshipType) as types
                    """).single()['types']
                }
                
                logger.info("Collected Neo4j stats")
                return stats
                
        except Exception as e:
            logger.error(f"Stats error: {e}")
            raise
    
    def close(self):
        """Close Neo4j connection."""
        if self.driver:
            self.driver.close()
            logger.info("Neo4j connection closed")

def main():
    """Main function."""
    graph = Neo4jGraph()
    
    try:
        # Example: Create nodes
        graph.create_node(
            'Person',
            {
                'name': 'John',
                'age': 30,
                'city': 'New York'
            },
            constraints=['name']
        )
        
        graph.create_node(
            'Person',
            {
                'name': 'Jane',
                'age': 28,
                'city': 'Boston'
            }
        )
        
        # Example: Create relationship
        graph.create_relationship(
            'Person',
            'KNOWS',
            'Person',
            {'since': '2023'},
            where={'name': 'John'}
        )
        
        # Example: Run PageRank
        results = graph.run_algorithm(
            'pagerank',
            {'nodeLabel': 'Person'}
        )
        
        # Example: Find paths
        paths = graph.query_paths(
            'Person',
            'Person',
            'KNOWS',
            max_depth=2
        )
        
        # Example: Monitor stats
        stats = graph.monitor_stats()
        
    finally:
        graph.close()

if __name__ == '__main__':
    main() 